# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Poultry Hero API
  version: 0.1.0
  description: SaaS Poultry Trading ERP — backend foundation (Phase 0/1).
paths:
...
  /api/v1/tenant/customers/{customer_id}/sales/:
    get:
      operationId: tenant_customers_sales_retrieve
      description: GET /api/v1/tenant/customers/{customer_id}/sales/
      parameters:
      - in: path
        name: customer_id
        schema:
          type: integer
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/tenant/sales/{id}/adjustments/:
    get:
      operationId: tenant_sales_adjustments_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
    post:
      operationId: tenant_sales_adjustments_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
...
  /api/v1/tenant/sales/{id}/collection-adjustment/:
    post:
      operationId: tenant_sales_collection_adjustment_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
  /api/v1/tenant/sales/{id}/lines/:
    get:
      operationId: tenant_sales_lines_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
    post:
      operationId: tenant_sales_lines_create
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SalesInvoiceDetail'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
...
  /api/v1/tenant/sales/{id}/print-preview/:
    get:
      operationId: tenant_sales_print_preview_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this sales invoice.
        required: true
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
  /api/v1/tenant/sales/price-preview/:
    get:
      operationId: tenant_sales_price_preview_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
  /api/v1/tenant/sales/stock-check/:
    get:
      operationId: tenant_sales_stock_check_retrieve
      description: |-
        Base viewset for tenant-owned resources.

        Subclasses set ``queryset``, ``permission_map`` (``{action: "perm.code"}``)
        and serializer(s). All rows are isolated by ``company_id``.
      tags:
      - tenant
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceDetail'
          description: ''
...
components:
  schemas:
...
    LineTypeEnum:
      enum:
      - product
      - by_product
      - service
      - other
      type: string
      description: |-
        * `product` - Product
        * `by_product` - By-product
        * `service` - Service
        * `other` - Other
...
    PaymentMethodC2cEnum:
      enum:
      - cash
      - bank_transfer
      - cheque
      - credit
      - other
      type: string
      description: |-
        * `cash` - Cash
        * `bank_transfer` - Bank Transfer
        * `cheque` - Cheque
        * `credit` - Credit / On Account
        * `other` - Other
...
    PaymentStatusEnum:
      enum:
      - unpaid
      - partially_paid
      - paid
      type: string
      description: |-
        * `unpaid` - Unpaid
        * `partially_paid` - Partially Paid
        * `paid` - Paid
...
    PriceSourceEnum:
      enum:
      - default_product_price
      - customer_special_price
      - manual_override
      - free_product
      type: string
      description: |-
        * `default_product_price` - Default Product Price
        * `customer_special_price` - Customer Special Price
        * `manual_override` - Manual Override
        * `free_product` - Free Product
    PriceTypeEnum:
      enum:
      - kg
      - piece
      - carton
      - tray
      type: string
      description: |-
        * `kg` - Per KG
        * `piece` - Per Piece
        * `carton` - Per Carton
        * `tray` - Per Tray
...
    SalesInvoiceAdjustment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        adjustment_type:
          $ref: '#/components/schemas/SalesInvoiceAdjustmentAdjustmentTypeEnum'
        title:
          type: string
          maxLength: 255
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        effect:
          $ref: '#/components/schemas/SalesInvoiceAdjustmentEffectEnum'
        reason:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - adjustment_type
      - created_at
      - effect
      - id
      - title
    SalesInvoiceAdjustmentAdjustmentTypeEnum:
      enum:
      - line_discount
      - invoice_discount
      - collection_adjustment
      - damaged_goods_adjustment
      - customer_settlement
      - commercial_discount
      - other
      type: string
      description: |-
        * `line_discount` - Line Discount
        * `invoice_discount` - Invoice Discount
        * `collection_adjustment` - Collection Adjustment
        * `damaged_goods_adjustment` - Damaged Goods Adjustment
        * `customer_settlement` - Customer Settlement
        * `commercial_discount` - Commercial Discount
        * `other` - Other
    SalesInvoiceAdjustmentEffectEnum:
      enum:
      - reduce_invoice_total
      - reduce_customer_balance
      - no_financial_effect
      type: string
      description: |-
        * `reduce_invoice_total` - Reduce Invoice Total
        * `reduce_customer_balance` - Reduce Customer Balance
        * `no_financial_effect` - No Financial Effect
    SalesInvoiceDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        invoice_number:
          type: string
          maxLength: 64
        customer:
          type: integer
        customer_name_snapshot:
          type: string
          maxLength: 255
        customer_trn_snapshot:
          type: string
          maxLength: 32
        customer_phone_snapshot:
          type: string
          maxLength: 32
        customer_address_snapshot:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusCd9Enum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        payment_method:
          $ref: '#/components/schemas/PaymentMethodC2cEnum'
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        discount_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        taxable_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        vat_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        vat_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        balance_due:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        fifo_cost_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        gross_profit:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        posted_receivable:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        credit_limit_snapshot:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        credit_limit_override_used:
          type: boolean
        credit_limit_override_reason:
          type: string
        vat_enabled:
          type: boolean
          readOnly: true
        notes:
          type: string
        approval_reason:
          type: string
        approved_by:
          type: integer
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        cancel_reason:
          type: string
        cancelled_by:
          type: integer
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        created_by:
          type: integer
          nullable: true
        updated_by:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/SalesInvoiceLine'
          readOnly: true
        adjustments:
          type: array
          items:
            $ref: '#/components/schemas/SalesInvoiceAdjustment'
          readOnly: true
      required:
      - adjustments
      - created_at
      - customer
      - id
      - invoice_date
      - invoice_number
      - lines
      - updated_at
      - vat_enabled
    SalesInvoiceLine:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
          nullable: true
        product_name_snapshot:
          type: string
          readOnly: true
        product_sku_snapshot:
          type: string
          readOnly: true
        line_type:
          $ref: '#/components/schemas/LineTypeEnum'
        quantity_cartons:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        quantity_pieces:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        quantity_kg:
          type: string
          format: decimal
          pattern: ^-?\d{0,11}(?:\.\d{0,3})?$
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        price_type:
          $ref: '#/components/schemas/PriceTypeEnum'
        price_source:
          $ref: '#/components/schemas/PriceSourceEnum'
        is_free:
          type: boolean
        free_reason:
          type: string
        line_subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
        taxable_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        vat_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        vat_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        line_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        fifo_cost_consumed:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        gross_profit:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        notes:
          type: string
        sort_order:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        is_stock_tracked:
          type: boolean
          readOnly: true
      required:
      - fifo_cost_consumed
      - gross_profit
      - id
      - is_stock_tracked
      - line_subtotal
      - line_total
      - product_name_snapshot
      - product_sku_snapshot
      - taxable_amount
      - vat_amount
...
    StatusCd9Enum:
      enum:
      - draft
      - approved
      - partially_paid
      - paid
      - cancelled
      type: string
      description: |-
        * `draft` - Draft
        * `approved` - Approved
        * `partially_paid` - Partially Paid
        * `paid` - Paid
        * `cancelled` - Cancelled
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT